+++ src/lib/constants.ts
@@ -1,3 +1,20 @@
+import {ROUTES} from '@/lib/routes'
+
+export const HEADER_MENU = [
+  {
+    label: 'Favorites',
+    to: ROUTES.favorites,
+  },
+  {
+    label: 'About',
+    href: 'https://snable.website',
+  },
+  {
+    label: 'Rate us',
+    href: 'https://snable.website/reviews',
+  },
+]
+
 export const FONT_WEIGHTS: Record<string, string> = {
   '100': 'thin',
   '200': 'ultralight',
@@ -8,4 +25,9 @@ export const FONT_WEIGHTS: Record<string, string> = {
   '700': 'bold',
   '800': 'extrabold',
   '900': 'black',
+}
+
+export const MODULE_STYLE = {
+  box: 'size-[42px] bg-control rounded-lg grid place-items-center',
+  icon: 'text-gray group-hover:text-white group-hover:scale-[1.05] duration-300',
 }