+++ vite.config.ts
@@ -31,8 +31,10 @@ export default defineConfig(({mode}) => {
     resolve: {
       alias: {
         '@': resolve(__dirname, './src'),
+        _hooks: resolve(__dirname, './src/hooks/'),
         _scripts: resolve(__dirname, './src/contentScript/'),
         _bg: resolve(__dirname, './src/background/'),
+        _modules: resolve(__dirname, './src/modules/'),
         '~': resolve(__dirname, './src/components'),
         '~~popup': resolve(__dirname, './src/components/modules/popup'),
         $: resolve(__dirname, './src/assets/'),