+++ CHANGELOG.md
@@ -1,5 +1,58 @@
 # CHANGELOG
 
+## 0.1.7
+
+### Patch Changes
+
+- 72ddeb6: add `hasFavorites` method for `favouritesManager`
+- 4c9d86c: add `onRemove` prop for `TabData` in `favorites` view
+- 427ddb4: add `view` prop for `TabData` Header sub-component
+- 048502e: upgrade `ExpandButton` UI sub-component
+- 4a05cbf: init `/favorites` popup page
+- f130a13: init `Dropdown` UI component based on `shadcn/ui`
+- 57390e6: decompose `Header` into `Logo` and `TabInfo` components
+
+## 0.1.6
+
+### Patch Changes
+
+- a1d1e7d: init `favouritesManager` file for `Header` Global component
+- de0a05a: init `favorites` button in `Header` Global component
+
+## 0.1.5
+
+### Patch Changes
+
+- 71ab634: upgrade `Header` Global component
+
+## 0.1.4
+
+### Patch Changes
+
+- b68cdfc: init `imagesExtractor` and `ImagesUnit` for its data
+- 5b3ba4b: init `Gallery` images block
+- 411e427: init `ExpandButton` UI sub-component (Button)
+- 62ebe72: init `ImageCell` UI component
+- 050170f: init `/images` popup sub-page
+
+## 0.1.3
+
+### Patch Changes
+
+- 3fc58c2: make `Controls` colors block
+- cf354d4: init `useCopy` custom hook
+- 22bf9f0: disable `scrollbar` in extension interface
+- 159f4f6: init `ColorCell` UI component
+- 040cf5e: init `Palette` colors block
+- 88ec5b9: init `/colors` popup sub-page
+
+## 0.1.2
+
+### Patch Changes
+
+- 02b36c7: enhance `Button` with `wouter` routing
+- 0706705: configure `routing` using `wouter`
+
 ## 0.1.1
 
 ### Patch Changes
@@ -25,7 +78,6 @@
 
 ### Patch Changes
 
-- 2e08f62: finalize `Header` Global component
 - ec173f9: pass `GET_TAB_INFO` data into `Header` Global component
 - d9eb5ec: init `getTabInfo` background worker for parsing `activeTab` data